1) users.php - change edit functions and toggling - added highlight to indicate editable rows - added dialogue to confirm actions

+++ Typemaster/php/deleteUser.php
@@ -4,11 +4,11 @@ require_once("db_connection.php");
 
 $conn = connection();
 
-
+if ($_SERVER["REQUEST_METHOD"] == "POST") {
     $id = $_POST['id'];
 
     $sql = "DELETE FROM users WHERE id = '$id'";
     $rs = mysqli_query($conn, $sql);
-
+}
 
 $conn->close();
